requirements.txt
requests>=2.25.1
pyyaml>=5.4.1
flask>=2.0.0
flask-restful>=0.3.9
flask-cors>=3.0.10
flasgger>=0.9.5
marshmallow>=3.14.0
watchdog>=2.1.0
python-dotenv>=0.19.0
